net: ipa: Add IPA v2.6L register definitions
IPA v2.6L is an older version the IPA hardware. It is the last 32 bit
IPA. Most of the registers were just shifted in newer IPA versions, but
the register fields have remained the same across IPA versions. This
means that only the register addresses needed to be added to the driver.

To handle the different IPA register addresses, static inline functions
have been defined that return the correct register address.

Signed-off-by: Sireesh Kodali <scarface@disroot.org>